requirements.txt
# RAG vs Raw LLM Demo - Dependencies
# Install with: pip install -r requirements.txt

# Web requests and HTML parsing
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Vector store and numerical computation
faiss-cpu>=1.7.4
numpy>=1.24.0

# Embeddings (local, free)
sentence-transformers>=2.2.0

# LLM APIs
anthropic>=0.18.0
openai>=1.0.0  # Optional, for OpenAI embeddings

# Web UI
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
pandas>=2.0.0
